pyproject.toml
[tool.poetry]
name = "GAME"
version = "1.2.047"
description = "GAME (Goals And Motivation Engine)"
authors = ["fvergaracl <fvergaracl@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
aiosqlite = "0.19.0"
alembic = "1.13.1"
anyio = "3.6.1"
asgiref = "3.7.2"
asyncpg = "0.29.0"
attrs = "23.2.0"
bcrypt = "4.1.2"
certifi = "2024.2.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
coverage = "7.4.1"
cryptography = "44.0.1"
dependency-injector = "^4.46.0"
ecdsa = "0.18.0"
execnet = "2.0.2"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "0.16.3"
httptools = "0.6.1"
httpx = "0.23.3"
idna = "3.7"
importlib-metadata = "7.0.1"
importlib-resources = "6.1.1"
iniconfig = "2.0.0"
isort = "5.13.2"
loguru = "0.7.2"
mako = "1.3.2"
markupsafe = "2.1.4"
packaging = "23.2"
passlib = "1.7.4"
pluggy = "1.4.0"
psycopg2-binary = "2.9.9"
py = "1.11.0"
pyasn1 = "^0.4.8"
pycparser = "2.21"
pydantic = "1.10.13"
pymysql = "1.1.1"
pyparsing = "3.1.1"
pytest = "8.0.0"
pytest-asyncio = "0.19.0"
pytest-cov = "4.1.0"
pytest-forked = "1.6.0"
pytest-mock = "3.12.0"
pytest-xdist = "3.5.0"
python-dotenv = "1.0.1"
python-jose = "^3.4.0"
pytz = "2024.1"
pyyaml = "6.0.1"
requests = "2.32.2"
rfc3986 = "1.5.0"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "1.4.36"
sqlalchemy2-stubs = "0.0.2a38"
sqlmodel = "0.0.8"
tomli = "2.0.1"
typing-extensions = "4.9.0"
urllib3 = "2.2.2"
uvloop = "0.19.0"
watchfiles = "0.21.0"
watchgod = "0.8.2"
websockets = "12.0"
zipp = "3.17.0"
vulture = "^2.11"
toml = "^0.10.2"
eralchemy = "^1.2.10"
uvicorn = "^0.30.1"
prettytable = "^3.11.0"
fastapi = "^0.114.2"
starlette = "^0.38.5"
prometheus-fastapi-instrumentator = "^7.0.0"
freezegun = "^1.5.1"
graphviz = "^0.20.3"
pyjwt = "^2.10.1"
numpy = "^2.2.3"
sentry-sdk = { extras = ["fastapi"], version = "^2.22.0" }
sphinx-material = "^0.0.36"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"
sphinx-autobuild = "^2024.2.4"
commitizen = "^3.25.0"
flake8 = "^7.0.0"
black = "^24.4.2"
ruff = "^0.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
release = "cz.cli:main"
